+++ resources/components/laraknife/setposition.blade.php
@@ -1,0 +1,7 @@
+@props(['position' => ''])
+@if($position === 'first')
+<div class="row">
+@endif
+@if($position === 'last')
+</div>
+@endif