+++ resources/components/laraknife/sortable-table-panel.blade.php
@@ -2,7 +2,7 @@
 <div class="lkn-form-table">
     <x-laraknife.hidden-button />
     <input name="_sortParams" id="_sortParams" type="hidden" value="{{ $fields['_sortParams'] }}">
-    <input type="hidden" name="pageIndex" id="_pageIndex" value="$pagination->pageIndex">
+    <input type="hidden" name="pageIndex" id="_pageIndex" value="{{ $pagination->pageIndex }}">
     <div class="lkn-pagination-block">
         <ul class="pagination lkn-pagination">
             @php $list = $pagination->listItems();  @endphp